conan: Allow for custom index in remote
By default we will just add the emzeat remote to the list of remotes
so packages can be found. Unless it is overwritten from the env.

The supported env variables can be used to change the order of remotes
so that binary packages found on the emzeat remote get preferred and
we remove hence load from conancenter.io