linting: Fix regression introduced by 2b8cfeb
When forcing ccache-tidy.py to go through python3 also on Windows we accidentially
broke the ccache acceleration because ccache was not able to parse the commandline
anymore. This is because it expects a single compiler executable as first argument only.

As an alternative solution there is CCACHE_PREFIX which we now use to pass the
python3 executable needed to run the python file.