cache-tidy: Rename from ccache-tidy
A new feature released with ccache 4.7 allows for ccache to masquerade
as the compiler by simply copying the ccache binary. As part of this
a detection was added to determine the location of the ccache binary
in the path which is simply testing if the program name starts with
the 'ccache' string.

This falsely triggered for the name ccache-tidy breaking compiler
detection and causing ccache to bail out claiming '-c' is not in the
PATH.

Solve by removing the ccache string from the script name avoiding this
and future naming collisions.

For readability reasons only the logging prefix and env variables
consumed by the script have been adapted, variable names used within
the script remain unchanged.

Reference commit in ccache:
	https://github.com/ccache/ccache/commit/ \
			451d92c801d6463bcc54c090b871a1d77f26d9d4